% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohviz.R
\name{compare_texts}
\alias{compare_texts}
\title{Scatterplot of word frequencies and 
match between two texts}
\usage{
compare_texts(tidy_text_one, tidy_text_two)
}
\description{
Scatterplot of word frequencies and 
match between two texts
}
